+++ engine/components/BackgroundComponent/BackgroundComponent.cpp
@@ -7,7 +7,7 @@ BackgroundComponent::BackgroundComponent(const char *filename) {
   strcat(this->path, filename);
 }
 
-BackgroundComponent::BackgroundComponent(const char *filename, double z) {
+BackgroundComponent::BackgroundComponent(const char *filename, float z) {
   strcat(this->path, "../assets/");
   strcat(this->path, filename);
   this->z = z;
@@ -42,18 +42,17 @@ void BackgroundComponent::loop() {
   auto size = gameObject->getSize();
   auto windowSize = gameObject->getGame()->getWindowSize();
 
-  int xStartOffset = (position.x / size.width + 1) * -1;
-  int xEndOffset = ((windowSize.width - position.x) / size.width + 1);
+  int xStartOffset = (int) (position.x / size.width + 1) * -1;
+  int xEndOffset = (int) ((windowSize.width - position.x) / size.width + 1);
 
   for (int x = xStartOffset; x < xEndOffset; x++) {
 
-    SDL_Rect rect{(int) (position.x + size.width * x), (int) (position.y * z * 0.5) - 150, static_cast<int>(size.width + 1.2),
-                  static_cast<int>(size.height + 1.2)};
+    SDL_Rect rect{(int) (position.x + size.width * (float) x), (int) (position.y * z * 0.5) - 150, (int) (size.width + 1.2),
+                  (int) (size.height + 1.2)};
 
     SDL_RenderCopy(getRenderer(), texture, nullptr, &rect);
 
   }
-
 
 }
